requirements.txt
pytorch
tensorly
hyperopt
matplotlib
pandas
dask
dask-ml
lightgbm
